Make FF_Router_Trips dispensable - Move functionality to FF_Router - Move mesh into ff_router as it is only used there - Update demo accordingly
